fixed lfc retrieval
level of free convection can have many names in the HRRR files (that's what I've obsevred on AWS like in some files its named no_level and others its named reserved) and I have no clue why this is but it is what it is. code accounts for this properly now. also attempted to fix a longstanding issue with hrrr cdo.rempann randomly not working and saying netcdf id not found